Move DTO to dto/ subdir
