+++ rest-client-quickstart/src/main/java/org/csdg8/user/UserResource.java
@@ -3,7 +3,7 @@ package org.csdg8.user;
 import org.csdg8.model.exception.InvalidCredentialsException;
 import org.csdg8.model.exception.UserAlreadyExistsException;
 import org.csdg8.model.exception.UserNotFoundException;
-import org.csdg8.user.UserController.RegistrationRequest;
+import org.csdg8.user.dto.RegistrationRequest;
 import org.eclipse.microprofile.openapi.annotations.Operation;
 import org.eclipse.microprofile.openapi.annotations.responses.APIResponse;
 import org.jboss.resteasy.reactive.RestResponse;
