+++ rest-client-quickstart/src/main/java/org/csdg8/user/dto/RegistrationRequest.java
@@ -1,0 +1,11 @@
+package org.csdg8.user.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.NoArgsConstructor;
+
+@AllArgsConstructor
+@NoArgsConstructor
+public class RegistrationRequest {
+    public String username;
+    public String password;
+}
