+++ rest-client-quickstart/src/main/java/org/csdg8/user/dto/UserResponse.java
@@ -1,0 +1,14 @@
+package org.csdg8.user.dto;
+
+import java.util.Set;
+
+import lombok.AllArgsConstructor;
+import lombok.NoArgsConstructor;
+
+@AllArgsConstructor
+@NoArgsConstructor
+public class UserResponse {
+    public Long id;
+    public String username;
+    public Set<String> role;
+}
